requirements.txt
python-escpos>=3.1
escpos-python
Pillow
requests
ijson
